chore(types): move centralized api types to domain-specific packages
