+++ pkg/api/vulnerability_handler.go
@@ -25,6 +25,23 @@ import (
 	"github.com/goharbor/harbor-cli/pkg/utils"
 )
 
+type ListVulnerabilityOptions struct {
+	CVEID       string
+	CVSSScore   string
+	Severity    string
+	Repository  string
+	ProjectName string
+	Package     string
+	Tag         string
+	Digest      string
+	Exclude     string
+	WithTag     bool
+	Fixable     bool
+	All         bool
+	Page        int64
+	PageSize    int64
+}
+
 func GetVulnerabilitySummary(withDangerousArtifact, withDangerousCVE bool) (*securityhub.GetSecuritySummaryOK, error) {
 	ctx, client, err := utils.ContextWithClient()
 	if err != nil {
